parsetags and gsps tool updates
