+++ docs/examples/GSPSSupport.html
@@ -208,7 +208,7 @@
       </div>
     </div>
 
-    <script src="./larvitar.js"></script>
+    <script src="../../dist/larvitar.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>
 
     <script>
@@ -280,7 +280,7 @@
           .catch(err => console.error(err));
       }
 
-      let demoFileList = ["test52.dcm", "test52.pres"];
+      let demoFileList = ["test208.dcm", "test208New.pres"];
       _.each(demoFileList, function (demoFile) {
         createFile(demoFile, renderSerie);
       });
