+++ imaging/imageTags.ts
@@ -488,7 +488,8 @@ export function parseTag<T>(
     } else if (
       propertyName === "x00700010" ||
       propertyName === "x00700011" ||
-      propertyName === "x00700014"
+      propertyName === "x00700014" ||
+      propertyName === "x00700273"
     ) {
       let pointsCoords = [];
       for (let index = 0; index < 2; index++) {
